# client/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API",
    "version" : "v1",
    "title" : "API"
  },
  "basePath" : "/",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/api/v1/functions" : {
      "get" : {
        "tags" : [ "Functions" ],
        "summary" : "list",
        "description" : "",
        "operationId" : "list",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/KFFunction"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Functions" ],
        "summary" : "create",
        "description" : "",
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CreateFunctionRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/KFFunction"
            }
          }
        }
      }
    },
    "/api/v1/functions/{id}" : {
      "get" : {
        "tags" : [ "Functions" ],
        "summary" : "get",
        "description" : "",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/KFFunction"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Functions" ],
        "summary" : "delete",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/api/v1/functions/{id}/proxy" : {
      "post" : {
        "tags" : [ "Functions" ],
        "summary" : "proxy",
        "description" : "",
        "operationId" : "proxy",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ProxyResponse"
            }
          }
        }
      }
    },
    "/api/v1/functions/{id}/service" : {
      "get" : {
        "tags" : [ "Functions" ],
        "summary" : "getAddress",
        "description" : "",
        "operationId" : "getService",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          }
        }
      }
    },
    "/api/v1/functions/{id}/versions" : {
      "get" : {
        "tags" : [ "Functions" ],
        "summary" : "getVersions",
        "description" : "",
        "operationId" : "getVersions",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Version"
              }
            }
          }
        }
      }
    },
    "/api/v1/functions/{id}/versions/{versionId}" : {
      "get" : {
        "tags" : [ "Functions" ],
        "summary" : "getVersion",
        "description" : "",
        "operationId" : "getVersion",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "versionId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Version"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Functions" ],
        "summary" : "setVersion",
        "description" : "",
        "operationId" : "setVersion",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "versionId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "CreateFunctionRequest" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "generic", "nodejs", "static_site" ]
        }
      }
    },
...
    "IntOrString" : {
      "type" : "object",
      "properties" : {
        "IntVal" : {
          "type" : "integer",
          "format" : "int32"
        },
        "Kind" : {
          "type" : "integer",
          "format" : "int32"
        },
        "StrVal" : {
          "type" : "string"
        }
      }
    },
    "KFFunction" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "owner" : {
          "$ref" : "#/definitions/UserNamespace"
        },
        "name" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "generic", "nodejs", "static_site" ]
        },
        "gitUrl" : {
          "type" : "string"
        },
        "namespace" : {
          "type" : "string"
        },
        "service" : {
          "type" : "string"
        },
        "deployment" : {
          "type" : "string"
        },
        "commitId" : {
          "type" : "string"
        },
        "ingressUrl" : {
          "type" : "string"
        }
      }
    },
...
    "LoadBalancerIngress" : {
      "type" : "object",
      "properties" : {
        "hostname" : {
          "type" : "string"
        },
        "ip" : {
          "type" : "string"
        }
      }
    },
    "LoadBalancerStatus" : {
      "type" : "object",
      "properties" : {
        "ingress" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LoadBalancerIngress"
          }
        }
      }
    },
...
    "ObjectMeta" : {
      "type" : "object",
      "properties" : {
        "annotations" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "clusterName" : {
          "type" : "string"
        },
        "creationTimestamp" : {
          "type" : "string"
        },
        "deletionGracePeriodSeconds" : {
          "type" : "integer",
          "format" : "int64"
        },
        "deletionTimestamp" : {
          "type" : "string"
        },
        "finalizers" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "generateName" : {
          "type" : "string"
        },
        "generation" : {
          "type" : "integer",
          "format" : "int64"
        },
        "labels" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "name" : {
          "type" : "string"
        },
        "namespace" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 253,
          "pattern" : "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
        },
        "ownerReferences" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OwnerReference"
          }
        },
        "resourceVersion" : {
          "type" : "string"
        },
        "selfLink" : {
          "type" : "string"
        },
        "uid" : {
          "type" : "string"
        }
      }
    },
    "OwnerReference" : {
      "type" : "object",
      "properties" : {
        "apiVersion" : {
          "type" : "string"
        },
        "kind" : {
          "type" : "string"
        },
        "controller" : {
          "type" : "boolean",
          "default" : false
        },
        "name" : {
          "type" : "string"
        },
        "uid" : {
          "type" : "string"
        }
      }
    },
    "ProxyResponse" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "body" : {
          "type" : "string"
        }
      }
    },
...
    "Service" : {
      "type" : "object",
      "required" : [ "apiVersion", "kind" ],
      "properties" : {
        "apiVersion" : {
          "type" : "string"
        },
        "kind" : {
          "type" : "string"
        },
        "metadata" : {
          "$ref" : "#/definitions/ObjectMeta"
        },
        "spec" : {
          "$ref" : "#/definitions/ServiceSpec"
        },
        "status" : {
          "$ref" : "#/definitions/ServiceStatus"
        }
      }
    },
    "ServicePort" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 63,
          "pattern" : "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$"
        },
        "nodePort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "port" : {
          "type" : "integer",
          "format" : "int32"
        },
        "protocol" : {
          "type" : "string"
        },
        "targetPort" : {
          "$ref" : "#/definitions/IntOrString"
        }
      }
    },
    "ServiceSpec" : {
      "type" : "object",
      "properties" : {
        "clusterIP" : {
          "type" : "string"
        },
        "deprecatedPublicIPs" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "externalIPs" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "externalName" : {
          "type" : "string"
        },
        "loadBalancerIP" : {
          "type" : "string"
        },
        "loadBalancerSourceRanges" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "ports" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServicePort"
          }
        },
        "selector" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "sessionAffinity" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "ServiceStatus" : {
      "type" : "object",
      "properties" : {
        "loadBalancer" : {
          "$ref" : "#/definitions/LoadBalancerStatus"
        }
      }
    },
...
    "UserNamespace" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "gitUser" : {
          "type" : "string"
        },
        "gitPassword" : {
          "type" : "string"
        }
      }
    },
...
    "Version" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        },
        "author" : {
          "type" : "string"
        },
        "timestamp" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    }
  }
}